refactor(json-schema): :fire: killed json schema package
