+++ packages/entities/package.json
@@ -9,7 +9,7 @@
     "access": "public"
   },
   "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "types": "./types/index.ts",
   "files": [
     "/dist"
   ],
@@ -18,7 +18,7 @@
     "url": "git+https://github.com/IQ-tech/firebolt.git"
   },
   "scripts": {
-    "generate-schema": "typescript-json-schema types-test.ts IExperienceJSONSchema --ignoreErrors --out test.json --jsdoc",
+    "generate-schema": "typescript-json-schema types/ExperienceConfig.ts IExperienceConfig --ignoreErrors --out schemas/experience-config.json --jsdoc",
     "clear-dist": "rimraf dist",
     "test": "jest",
     "test:coverage": "jest --collect-coverage",
