--- packages/entities/types/ExperienceConfigJSON/index.ts
@@ -1,11 +1,0 @@
-export interface IExperienceJSONSchema {
-  /** test */
-  "$schema-version"?: string
-  "$experience-version"?: string
-  name: string
-  description: string
-  business: string // todo remove
-  decisionHandlerConfig?: IDecisionHandlerConfig
-  flows: IFlow[]
-  steps: IStepJSON[]
-}
