+++ packages/entities/types/index.ts
@@ -1,0 +1,5 @@
+export interface IStepFormPayload {
+  [key: string]: string
+}
+
+export * from "./ExperienceConfig"
