--- packages/json-schema/CHANGELOG.md
@@ -1,16 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [0.7.0](https://github.com/IQ-tech/firebolt/compare/v0.6.0...v0.7.0) (2022-04-26)
-
-**Note:** Version bump only for package @iq-firebolt/json-schema
-
-
-
-
-
-# [0.6.0](https://github.com/IQ-tech/firebolt/compare/v0.5.0...v0.6.0) (2022-03-15)
-
-**Note:** Version bump only for package @iq-firebolt/json-schema
