--- packages/json-schema/README.md
@@ -1,3 +1,0 @@
-# `json-schema`
-
-package to validate firebolt json forms
