--- packages/json-schema/admin/vite.config.js
@@ -1,15 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-
-export default defineConfig({
-  root: "admin/ui",
-  publicDir: "admin/public/",
-  plugins: [
-    react({
-      jsxImportSource: "@emotion/react",
-      babel: {
-        plugins: ["@emotion/babel-plugin"],
-      },
-    }),
-  ],
-});
